Migrate inverseSqrt tests
